Move existing measurement error mitigation to legacy folder
